bitstream_files
I was able to abstract all the variations of types of fields create
independent huffman stats, along with optionally configurable hash
tables, log files, etc.

it also handles unclean shutdowns, at least mostly by having a block
num at the end of the file, so if they are all matching for all the
files then it is a good set. I am not buffering, so there are some
fields that are doing the current block before the batch is flushed to
disk, so need to recalibrate that after the block chain is recreated in
ram during init or to buffer the current block.

the stats are being generated on the fly, but unused as the time to
generate the data is a long time, especially for btc. Uncompressed log
files are being generated and while that is going, I can debug and
optimize the compression code, along with allowing for parallel
generation, etc.